+++ CHANGELOG.md
@@ -1,4 +1,11 @@
+
 # CHANGE LOG
+## v 0.2.1
+- lein -> CLI
+  - added build.clj and Makefile
+## v 0.1.3
+- Tweaks to lstr tests
+- Updating dependencies
 ## V 0.1.2 
 - Uses the new vocabulary format module